FormGUI nach dem MVC-Pattern aufgeteilt
