+++ src/de/muenchen/allg/itd51/wollmux/document/DocumentManager.java
@@ -61,9 +61,9 @@ import de.muenchen.allg.itd51.wollmux.core.parser.ConfigThingy;
 import de.muenchen.allg.itd51.wollmux.core.parser.NodeNotFoundException;
 import de.muenchen.allg.itd51.wollmux.core.util.L;
 import de.muenchen.allg.itd51.wollmux.core.util.Utils;
-import de.muenchen.allg.itd51.wollmux.dialog.formmodel.SingleDocumentFormModel;
 import de.muenchen.allg.itd51.wollmux.dialog.mailmerge.MailMergeNew;
 import de.muenchen.allg.itd51.wollmux.event.WollMuxEventHandler;
+import de.muenchen.allg.itd51.wollmux.form.control.FormController;
 import de.muenchen.allg.itd51.wollmux.former.FormularMax4kController;
 
 /**
@@ -85,9 +85,9 @@ public class DocumentManager
   private Map<HashableComponent, Info> info =
     new HashMap<>();
 
-  private Map<XTextDocument, SingleDocumentFormModel> formModels = new HashMap<>();
   private Map<XTextDocument, FormularMax4kController> fm4k = new HashMap<>();
   private Map<XTextDocument, MailMergeNew> mailMerge = new HashMap<>();
+  private Map<XTextDocument, FormController> controller = new HashMap<>();
 
   /**
    * Enthält alle registrierten XEventListener, die bei Statusänderungen der
@@ -254,28 +254,27 @@ public class DocumentManager
   }
 
   /**
-   * Liefert die zu diesem Dokument zugehörige FormularGUI, falls dem
-   * TextDocumentModel die Existent einer FormGUI über setFormGUI(...) mitgeteilt
-   * wurde - andernfalls wird null zurück geliefert.
+   * Liefert die zu diesem Dokument zugehörige FormularGUI, falls dem TextDocumentModel die Existent
+   * einer FormGUI über setFormGUI(...) mitgeteilt wurde - andernfalls wird null zurück geliefert.
    *
    * @return Die FormularGUI des Formulardokuments oder null
    */
-  public synchronized SingleDocumentFormModel getFormModel(XTextDocument doc)
+  public synchronized FormController getFormModel(XTextDocument doc)
   {
-    return formModels.get(doc);
+    return controller.get(doc);
   }
 
   /**
-   * Gibt dem TextDocumentModel die Existent der FormularGUI formGUI bekannt und wird
-   * vom DocumentCommandInterpreter in der Methode processFormCommands() gestartet
-   * hat, falls das Dokument ein Formulardokument ist.
+   * Gibt dem TextDocumentModel die Existent der FormularGUI formGUI bekannt und wird vom
+   * DocumentCommandInterpreter in der Methode processFormCommands() gestartet hat, falls das
+   * Dokument ein Formulardokument ist.
    *
    * @param formGUI
    *          Die zu diesem Dokument zugehörige formGUI
    */
-  public synchronized void setFormModel(XTextDocument doc, SingleDocumentFormModel formModel)
+  public synchronized void setFormModel(XTextDocument doc, FormController formModel)
   {
-    this.formModels.put(doc, formModel);
+    this.controller.put(doc, formModel);
   }
 
   /**
@@ -485,11 +484,11 @@ public class DocumentManager
     }
     mailMerge.remove(doc);
 
-    if (formModels.containsKey(doc) && formModels.get(doc) != null)
+    if (controller.containsKey(doc) && controller.get(doc) != null)
     {
-      formModels.get(doc).closing(doc);
+      controller.get(doc).closing(doc);
     }
-    formModels.remove(doc);
+    controller.remove(doc);
   }
 
   /**
