+++ src/de/muenchen/allg/itd51/wollmux/event/handlers/OnSetWindowPosSizeSingleForm.java
@@ -1,6 +1,6 @@
 package de.muenchen.allg.itd51.wollmux.event.handlers;
 
-import de.muenchen.allg.itd51.wollmux.dialog.formmodel.SingleDocumentFormModel;
+import de.muenchen.allg.itd51.wollmux.form.control.FormController;
 
 public class OnSetWindowPosSizeSingleForm extends BasicEvent
 {
@@ -8,13 +8,12 @@ public class OnSetWindowPosSizeSingleForm extends BasicEvent
   private int docY;
   private int docWidth;
   private int docHeight;
-  private SingleDocumentFormModel m;
+  private FormController formModel;
 
-  public OnSetWindowPosSizeSingleForm(SingleDocumentFormModel m, int docX,
-      int docY,
-      int docWidth, int docHeight)
+  public OnSetWindowPosSizeSingleForm(FormController formModel, int docX,
+      int docY, int docWidth, int docHeight)
   {
-    this.m = m;
+    this.formModel = formModel;
     this.docX = docX;
     this.docY = docY;
     this.docWidth = docWidth;
@@ -24,6 +23,6 @@ public class OnSetWindowPosSizeSingleForm extends BasicEvent
   @Override
   protected void doit()
   {
-    m.setWindowPosSize(docX, docY, docWidth, docHeight);
+    formModel.setWindowPosSize(docX, docY, docWidth, docHeight);
   }
 }
